mmc: Allow consistent mmc/mmcblk numbering by aliases
As with gpio, uart and others, allow specifying the name_idx via the
aliases-node in the devicetree.

On embedded devices, there is often a combination of removable (e.g.
SD card) and non-removable mmc devices (e.g. eMMC).
Therefore the name_idx might change depending on
- host of removable device
- removable card present or not

This makes it difficult to hard code the root device, if it is on the
non-removable device. E.g. if SD card is present eMMC will be mmcblk1,
if SD card is not present at boot, eMMC will be mmcblk0.

If the aliases-node is not found, the driver will act as before.

The original patch is from here:
http://linux-kernel.2935.n7.nabble.com/PATCH-v2-0-4-Patches-to-allow-
consistent-mmc-mmcblk-numbering-w-device-tree-td1365050.html