watchdog: imx2_wdt: Add keepaliveoff parameter
Since commit faad5de0b104 ("watchdog: imx2_wdt: convert to watchdog core api")
the default behaviour of this watchdog driver has changed. Now
if the watchdog was activated in the bootloader, it is
retriggered with a keepalive timer in the Kernel.

On some embedded systems, however, it is necessary to monitor
the entire boot process until the main application is started.

By activating this parameter (imx2-wdt.keepaliveoff) the old
behaviour can be used again. So if watchdog and this parameter
are activated the system will restart if the watchdog isn't
retriggered in time by an user space application.

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>