watchdog: da9062: Add timeout values for operation with internal oscillator
Add timeout values for operation with internal oscillator. According to
the data sheet [1] the internal oscillator has 25kHz whereras the external
ocillator should have 32kHz. So calculate the timeout values for the
internal timeout values by formula:

    2^(15+TWDSCALE)
t = ---------------
         freq

The DA9061 is always driven by the internal ocillator, because there is no
possibility to connect a crystal. For the DA9062 the reg/field EN_32K/CRYSTAL
is read and the timeout values are set accordingly.

[1] https://www.dialog-semiconductor.com/sites/default/files/2021-09/DA9062_Datasheet_3v7.pdf

Upstream-Status: Inappropriate [Not for automotive high temperature operation]

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>