ARM: dts: imx6qdl-dhcom: Identify the PHY by ethernet-phy-id0007.c0f0
Identify the PHY by its compatible ID value. In some cases during
boot, the PHY needs to be reset to be accessible, but this is only
possible if the PHY is recognized. In that case, the automatic
detection of the PHY does not work and a static compatible ID
value is need.

Upstream-Status: Backport [cbcf2b40a7cf40da02e08e4485118a3b4fce54ce]

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>
Cc: Shawn Guo <shawnguo@kernel.org>
Cc: Fabio Estevam <festevam@gmail.com>
Cc: Marek Vasut <marex@denx.de>
Cc: NXP Linux Team <linux-imx@nxp.com>
Cc: kernel@dh-electronics.com
To: linux-arm-kernel@lists.infradead.org