Revert "drm/panel-simple: drop use of data-mapping property"
Without the data-mapping devicetree property my display won't
work properly. It is flickering, because the bus flags won't
be assigned without a defined bus format by the imx parallel
display driver. There was a discussion about the removal [1]
and an agreement that a better solution is needed, but it is
missing so far. So what would be the better approach?

[1] https://patchwork.freedesktop.org/patch/357659/?series=74705&rev=1

This reverts commit d021d751c14752a0266865700f6f212fab40a18c.

Upstream-Status: Inappropriate [Another solution must be found]

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>