ARM: dts: imx6ull-dhcor: Remove mmc aliases from DHCOR SoM
To be able to define mmc aliases correctly for a DHCOR board
remove mmc aliases from the DHCOR SoM layer.

All DHCOM SoM boards are based on the DHCOR SoM, as the DHCOR
SoM is soldered on top of the DHCOM SoM. As a result, remove
the mmc /delete-property/ entries from the aliases node in
the DHCOM SoM layer, because this is already removed in the
DHCOR SoM layer below.

Upstream-Status: Submitted [https://lore.kernel.org/lkml/20230407104943.5794-3-cniedermaier@dh-electronics.com]

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>