cpufreq: imx6q: Reduce fallback to 792MHz if 900MHz isn't in OPP on i.MX6ULL
For a 900MHz i.MX6ULL CPU, if the specified OPP doesn't
define an operating point above 792MHz, reduce the speed
grade to 792MHz. This is a simple way to avoid falling
back to a maximum frequency of 528MHz for this CPU, as
otherwise the 792MHz frequency would also be removed.

Upstream-Status: Pending

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>