cpufreq: imx6q: No warning output when disabling a non-existing frequency of the OPP
It is confusing if a warning is given for disabling a non-existent
frequency of the operating performance points (OPP). In this case
the function dev_pm_opp_disable() returns -ENODEV. Check the return
value and avoid the output of a warning in this case. Avoid code
duplication by using a separate function.

Upstream-Status: Pending

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>