ARM: dts: imx6ull-dhcom: Remove /omit-if-no-ref/ from node usdhc1-pwrseq
Remove /omit-if-no-ref/ from node usdhc1-pwrseq, because if the compile
flag -@ (include symbols) is used the node will always be there. In this
case, GPIO H is not released and therefore cannot be used. Therefore,
remove this node manually from the corresponding devicetree file and
don't rely on /omit-if-no-ref/.

Upstream-Status: Submitted

Signed-off-by: Christoph Niedermaier <cniedermaier@dh-electronics.com>